Make palette selection and hover colors customizable
This defines a new ColorPalette interface that can be set in the
PaletteViewer. The colors provided by this interface are then used in
the palette figure for e.g. the tool entries.

Clients can extend the DefaultColorPalette to define their own colors
and therefore change the look of the palette.