Make "button" colors used inside the palette configurable
This removes the following constants from PaletteColorUtil and reroutes
their references to the PaletteColorProvider:

- WIDGET_BACKGROUND -> getButton()
- WIDGET_NORMAL_SHADOW -> getButtonDarker()
- WIDGET_DARK_SHADOW -> getButtonDarkest()

The "mixed" getButtonDarker() colors are created via the new method
getButtonDarker(double), where the value describes the weight with which
the color from getButton() color is mixed with getButtonDarker().

In addition, hover color for the PinFigure is calculated via the new
method getHoverColor().

Both new methods are marked as final, as they are derived from other
colors.