Update all lists of listeners to use CopyOnWriteArrayList
This changes the list of listeners to a thread-safe implementation to
avoid the ConcurrentModificationException when modifying this list while
in the middle of handling an event.

Note that this also adapts 71f0c56301033bcf5ff5ba1261cdc2719d9bb391, in
which the List of ISelectionChangedListeners was changed to a
ListenerList, thus breaking API on the protected field.