Remove deprecation from getDistanceOrthogonal #21
getDistanceOrthogonal is used in several places in GEF. There is no
straight forward replacement. To reason for deprecation is precision of
subclasses. As there is only one subclass with higher precision an
improved version of that method is added there. Furthermore the impact
on precision will anyhow only happen if both sides are PrecisionPoints.

https://github.com/eclipse-gef/gef-classic/issues/21